Add code, tests, and documentation for the package
This is a mega commit that adds all the code
necessary for the initial version of the package.

The README and docs/ describe the system design
and usage guide.

Given that it was the first iteration of the
package, and things were reverted and updated
very often, it didn't make sense to preserve the
entire commit history, so it's squashed into this
one commit. Moving forwards, commits should most
definitely be smaller.